add r_drawflat_nonworldmodel, defaults to 0
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@2315 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/client/snd_dma.c
@@ -1622,7 +1622,7 @@ void S_RawAudio(int sourceid, qbyte *data, int speed, int samples, int channels,
 //		Con_Printf("Restarting raw stream\n");
 	}
 
-	speedfactor	= (float)speed/snd_speed;
+	speedfactor	= (double)speed/snd_speed;
 	outsamples = samples/speedfactor;
 
 	oldlength = s->sfxcache->length;
